+++ spec/iex/endpoints/income_spec.rb
@@ -48,9 +48,18 @@ describe IEX::Resources::Income do
     end
   end
 
+  context 'no result', vcr: { cassette_name: 'income/nsrgy' } do
+    subject do
+      client.income('nsrgy')
+    end
+    it 'returns empty array' do
+      expect(subject).to eq []
+    end
+  end
+
   context 'invalid symbol', vcr: { cassette_name: 'income/invalid' } do
     subject do
-      client.earnings('INVALID')
+      client.income('INVALID')
     end
     it 'fails with SymbolNotFoundError' do
       expect { subject }.to raise_error IEX::Errors::SymbolNotFoundError, 'Symbol INVALID Not Found'